+++ src/tobi.backfrak.de/cmd/gpsa/input_test.go
